pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6.3",
    "colorlog>=6.9.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "langchain-community>=0.3.22",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.14",
    "pydantic>=2.11.3",
    "pypdf>=5.4.0",
    "python-multipart>=0.0.20",
    "rank-llm>=0.21.0",
    "redis>=6.0.0",
    "sentence-transformers>=4.1.0",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
